Store grades as ints
Round to the nearest whole number when saving. Doesn't lose data when
saving grades to the DB now. Convert to a letter grade is the job of the
front end. The grade methods for SPF and DMARC must now return a number.
This number should be out of 100? Should I add max_score type column,
that idea seems dumb and complex. Maybe just force it be the
percentage expressed, seems like a better idea.