Improve error handling for SPF IPv4 and IPv6
Give better errors about value error issues and catch all errors. This
is to fix issues with host bits in the CIDR address. But trying to
fend off future bugs.

Added a new error for invaild ipv4 and ipv6 values in the enum and
database. The numbers are designed to fit around !40.

Signed-off-by: Matthew Burket <m@tthewburket.com>